Extend Kubewarden steps and fix curl issue
Separate in Step 2 the creation of the docker group into a new command block, because there where often issues with this cpmmand block.
Added the insecure flag to the curl command in Step17
Extend in Step 24 steps to undo NV proctection mode, because if the node is in protected mode we can't install Kubewarden.